Fixed Bugs and added LocationString

+++ Version_LATEST/src/main/java/com/acrylic/version_latest/Animations/AbstractAnimations.java
@@ -69,7 +69,9 @@ public abstract class AbstractAnimations {
 
     public void delete() {
         entity.remove();
-        holograms.delete();
+        if (holograms != null) {
+            holograms.delete();
+        }
     }
 
     public abstract void teleport(Location location);
